feat: task, comment... 기능 구현

+++ src/main/java/com/nhnacademy/gateway/controller/TagController.java
@@ -4,6 +4,7 @@ import com.nhnacademy.gateway.api.TagApiClient;
 import com.nhnacademy.gateway.dto.tag.TagDeleteRequest;
 import com.nhnacademy.gateway.dto.tag.TagCreateRequest;
 import com.nhnacademy.gateway.validation.ValidationSequence;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
@@ -11,15 +12,16 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequiredArgsConstructor
+@RequestMapping("/projects/{project-id}/tags")
 public class TagController {
 
     private final TagApiClient tagApiClient;
 
-    // TODO-R 공통 처리 (/projects/{project-id}/tags)
-    @PostMapping("/projects/{project-id}/tags") // TODO-Q 경로?
+    @PostMapping // TODO-Q 경로?
     public String tagCreate(@Validated(ValidationSequence.class) @ModelAttribute TagCreateRequest request,
                             BindingResult bindingResult,
                             @PathVariable("project-id") Long projectId) {
@@ -32,10 +34,16 @@ public class TagController {
 
         return "redirect:/projects/" + projectId;
     }
-
-    @PostMapping("/projects/{project-id}/tags/deactivate")
-    public String tagDelete(@ModelAttribute TagDeleteRequest request,
+// task-tag : deactivate
+    @PostMapping("/delete")
+    public String tagDelete(@Valid @ModelAttribute TagDeleteRequest request,
+                            BindingResult bindingResult,
                             @PathVariable("project-id") Long projectId) {
+        if(bindingResult.hasErrors()) {
+
+            return "project/projectModify";
+        }
+
         tagApiClient.deleteTags(projectId, request);
 
         return "redirect:/projects/" + projectId;
